style 예약확인 페이지 반응형 css수정
